\name{NlogQLimits}
\alias{NlogQLimits}
\title{
Plot limits of NlogQ
}
\description{
Produces a 3d rotatable plot using \pkg{rgl} which gives values of NlogQ for various values of N and k.
}
\usage{
NlogQLimits()
}
\details{
Produces a 3d rotatable plot using \pkg{rgl} which gives values of NlogQ for various values of N and k.
\dfn{N} = number of observations, \dfn{k} = numer of outliers proposed to be rejected.
}
\value{
A 3d plot with \pkg{rgl}.
}
\seealso{
See Also \code{\link{NlogQ}}
}
\examples{
\dontrun{NlogQLimits()}
}