\name{Peirce-package}
\alias{Peirce-package}
\alias{Peirce}
\docType{package}
\title{
\packageTitle{Peirce}
}
\description{
\packageDescription{Peirce}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Peirce}
\packageIndices{Peirce}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Peirce}

Maintainer: \packageMaintainer{Peirce}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
