% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeirceChara.R
\name{Peirce.char}
\alias{Peirce.char}
\title{Peirce sign function for inital classification of character data}
\usage{
Peirce.char(DF)
}
\arguments{
\item{DF}{object of class data.frame with dim [n,3]}
}
\value{
list with the index of the first duplicated row, if any, 
and duplicated rows, the identity of those rows and how many times
each row is duplicated, excludes rows that are unique
}
\description{
\code{Peirce.char} evalutaes a data.frame for duplicated rows that represent Peirce's triads
}
\examples{
Peirce.char(x)
"No duplicate rows"

head(twitter)

Peirce.char(twitter)
[[1]]   19

}

