% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeirceWhole.R
\name{whlPeirce}
\alias{whlPeirce}
\title{Peirce Function for whole numbers}
\usage{
whlPeirce(DF)
}
\arguments{
\item{DF}{object of class data.frame with 3 columns}
}
\value{
a data frame with the average distribution for each of 3 variables
}
\description{
\code{whlPeirce} returns a data frame with the average distribution
}

