\name{Peirce-package}
\alias{Peirce-package}
\alias{Peirce}
\docType{package}
\title{Data classification based on Peirce's semiosis}
\packageTitle{Peirce}
\description{The package Peirce applies sign theory or semiosis to 
    find triadic relationships among three variables consisting of 
    whole numbers or character data.
    }
\author{Erin Feichtinger
Maintainer: Erin Feichtinger <efeichtinger@mail.usf.edu>}
\references{Friedman, A. and E. Feichtinger. 2017. Peirce's sign theory as
    an open-source R package. Signs vol 8:pp. 1-24. In press}
\examples{
charPeirce(twitter)
[[1]]   19
}
